scripts/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
segmentation-models-pytorch>=0.3.0
timm>=0.9.0
albumentations>=1.3.0
opencv-python>=4.5.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0
captum>=0.4.0
scikit-learn>=1.0.0
Pillow>=8.0.0
wandb>=0.17.0